src/analytics/lambdas/requirements.txt
# boto - refer to https://docs.aws.amazon.com/lambda/latest/dg/lambda-runtimes.html
boto3==1.18.55
botocore==1.21.55

# pytest
pytest==6.2.5
pytest-forked==1.3.0
pytest-mock==3.6.1
pytest-xdist==2.4.0

requests==2.26.0

google-cloud-storage==1.42.3
cryptography==35.0.0
PyJWT==2.3.0
